+++ Changelog.md
@@ -106,6 +106,9 @@ and this project follows to [Ragnarök Versioning Convention](https://gist.githu
   - Dumpster
   - Sandbag
   - Duel Floodlight
+  - Single Floodlight
+  - Shelf Top
+  - Server Rack Top
 
 ### Removed
 
@@ -115,6 +118,7 @@ and this project follows to [Ragnarök Versioning Convention](https://gist.githu
 - Removed most of the compatibility layer
 - Removed FXAA
 - Removed Bloom
+- Deleted broken Hescobastion and UMP45GFL props  
 
 ### Optimization